requirements.txt
Jinja2>=3.1
pydantic>=2.7
PyYAML>=6.0
python-dateutil>=2.9
rapidfuzz>=3.9
requests>=2.32
pytest>=8.0
